Add support for self hosted fonts
Building on the work done by @visualfanatic fonts can now be hosted outside of the `/static` folder and will be copied across by the plugin. From my recollection this is what is suggested in the Gatsby docs. Updated README. Included the Gatsby reporter to throw a panic if file copy to `/public` fails.